app/requirements.txt
fastapi[all]
uvicorn
elasticsearch
python-dotenv
pinecone-client
aider-chat
aiohttp
slack_bolt
openai